修改了后台的签到导出,以及用户批量导入对学院信息的处理

+++ framework/src/main/java/com/jx/domain/vo/UserVolunteerRecordVo.java
@@ -12,5 +12,5 @@ public class UserVolunteerRecordVo {
     private String userName;
     private String departmentName;
     private Long userParticipateInActivityNum;
-    private Long userVolunteerTimes;
+    private Double userVolunteerTimes;
 }
